chg: [improvement] format output + typo const
