+++ ail_typo_squatting/generator/const/main.py
