+++ ail_typo_squatting/generator/subdomain.py
@@ -1,7 +1,6 @@
 # Import all the utils
 from .utils.generator_functions import *
 
-
 """
 
    Original Domain        Typosquatted Domain
